[compiler] Polymorph & evaluation cache.
- Redesign recipe instance cache to support also comptime arguments. This reduced count of generated polymorphs rapidly.
- Add comptime-call result cache, function call can be in some cases reused instead of calling the function again.
- Add #nocache directive to disable caching of comptime function calls.
- Reduce size of `mir_fn` struct using optional lazily allocated extensions.

+++ src/bldebug.h
@@ -109,7 +109,7 @@ void print_trace_impl(void);
 	(void)0
 #define bmagic_member                  void *_magic;
 #define bmagic_set(O)                  (O)->_magic = (void *)&(O)->_magic
-#define bcalled_once_member(name)      s32 _##name##_call_count;
+#define bcalled_once_member(name)      s8 _##name##_call_count;
 #define bcalled_once_assert(obj, name) bassert((obj)->_##name##_call_count++ == 0 && "Expected to be called only once!")
 #define bcheck_main_thread()           bassert(thrd_equal(MAIN_THREAD, thrd_current()) && "Function is supposed to be called from the main thread!")
 #define bcheck_true(expr)              bassert(expr)
